Augur was a single monolith, not two monoliths
Rewrites the background paragraph such that the monolithic single Augur
script is introduced early enough for other sentences to meaningfully
reference this script. As part of this rewriting, I've tried to clarify
what was confusing about the paragraph before Richard's edits.
Specifically, I've noted that the original script was part of nextflu
and only worked with flu data. The general flow of the paragraph is now
like this:

1. Augur as a monolith in nextflu
2. Augur as many monoliths in Nextstrain
3. Redundancy and confusion for all, as a result of many monoliths,
   warranting a modular Augur.