Prototype augur filter sequence index interface
Adds an optional `--sequence-index` argument to augur filter that
accepts the output of the new `augur index` command and uses these
sequence composition data to filter sequences by length or invalid
nucleotide composition instead of inspecting sequences directly.

This prototype interface is backwards-compatible with previous workflows
that don't precalculate the sequence index. As such, it tries to
calculate the index in the fly when one is not provided by the user.
This commit stubs out the functions in `index.py` that would allow this
on-the-fly calculation.

The proposed `index_sequences` function would contain the current body
of the `augur index` module's `run` function and call the proposed
`index_sequence` function on each input sequence. The separate
`index_sequence` allows us to test the most important logic of the index
module with doctests or unit tests. This commit provides some doctests
as examples of how this function could work.