Use probabilistic sampling by default
Makes the optional flag `--probabilistic-sampling` enabled by default
and adds a `--no-probabilistic-sampling` flag to disable this default
behavior. These changes make the `--probabilistic-sampling` flag
unnecessary but maintain backwards-compatibility for workflows that
reference this flag.

We toggle probabilistic sampling through one variable. This simplifies
the downstream logic that checks for whether to use probabilistic
sampling or not. Note that with this approach we cannot define a
separate `help` string for the `--no-probabilistic-sampling` flag
because it will appear on the command line with a default value of its
destination variable (`probabilistic_sampling=True`).

This commit also disables probabilistic sampling for test builds. Test
builds sample so few sequences that probabilistic subsampling can end up
picking too few sequences for augur refine to work properly. This commit
disables probabilistic sampling for those builds.

Finally, this commit removes the redundant Zika test build implemented
with Snakemake since we already have a Cram test (`zika.t`) that does
the same thing.