frequencies: Allow YYYY-MM-DD dates for min and max date
This commit copies approach of e67d3df724342097eade5c41cc9955f0dbad2969 to allow entry of --min-date and --max-date as ISO 8601 YYYY-MM-DD.

This commit also fixes float_to_datestring to work with output that are higher resolution than the first of the month. This copies the function used in treetime.utils called float_to_datestring to do so.