Print index statistics when explicitly requested
Proposes a default of _not_ printing index statistics to stdout unless
the user requests that information with the new `--verbose` flag. This
minor change allows users to take advantage of the fact that the index
is streamed to its output file handle, so they can pipe the output to
stdout and into other tools like gzip to produce compressed index files.