Intersect sequences and metadata strains during filter
When sequences input (or a sequence index) is provided, intersect those
strains with the strains in the metadata while also tracking the number
of strains filtered by this operation.

Note that this approach is not too different from the original filter
implementation that required each sequence to have metadata or be
filtered out. This approach has the additional effect of filtering out
strains that have metadata and no sequences which is a filter that used
to happen more implicitly before. This commit adds the number of strains
filtered by this operation to the report at the end, so users have a
better sense of which strains are missing data.

As a side effect of getting the functional tests in this commit to work,
this commit also reorders the report output such that the user always
sees how many strains were filtered and why before the error message
gets printed when all strains are filtered.

This commit also adds a functional test for filtering VCFs.

Fixes #681