Add tests and logic for inline comments in strain lists
Previously, augur filter treated comments differently for `include` and
`exclude` files. This commit consistently allows the more permissive
commenting associated with the `exclude` files such that users can make
inline comments on the same line as a strain name. This commit also adds
tests to confirm that we get the expected behavior from the current
code. This change should be backwards compatible with existing `include`
inputs, since the new comment standard is a superset of the older
`include` format.