Remove Zika and TB tutorials from docs
Tutorial content now lives in the docs.nextstrain.org repository, so
this commit removes the redundant content from these tutorials and
references to these tutorials. A copy of the Zika tutorial from this
repo will live in the docs.nextstrain.org repo where it will be
rewritten as a more complete introduction to converting the Augur-only
version of the Zika tutorial to an automated workflow with Snakemake.