Update strain counts after writing sequence output
It is possible for the sequence index to be a superset of (or generally
out of sync with) the given sequences output. To handle this case, we
track the observed sequence strains when the user requests sequence
output and update the statistics for which strains were filtered and
why. We also move the handling of sequence outputs before the metadata
and strain list outputs, so these outputs reflect any changes in the
available sequences that occur when writing sequence output.