Replace environment file with Bioconda packages
Replaces references to local Conda environment file with conda install
commands that install the Augur package from Bioconda. This approach was
not previously possible before we started maintaining the Augur Bioconda
package.

As part of these changes, this commit also improves the usability of the
`./devel/test` script a bit.

Fixes #421