Provide control over numeric format of distances
Provides two new top-level keys in the distance map definition,
`output_type` and `precision`, that allow the user to set the output
type to "integer" or "int" (overriding the default type of floating
point values) and set the precision of the floating point output.
Updates the doctests and documentation to match these new features.