Filter by sequences when they are provided
Fixes logic for whether to try to generate a sequence index, change the
warning about the auto-generated sequence index to a notice, and make
sure outputs are filtered by the provided sequences when no sequence
index is provided. Historically, we've filtered data such that we only
keep those strains that are present in the metadata and the sequence
data when sequence information is provided. With the implementation in
this commit, we skip all interaction with sequence data when no sequence
or sequence index paths are provided. When these paths are provided, we
use these data to filter strains even when the user has not requested
sequence output.