Filter VCFs by sequence index like sequences
Adds the ability to filter VCFs by a sequence index, to standardize the
filtering interface for both FASTA and VCF inputs. This works by adding
a new `index_vcf` function to the `index` module along with support for
indexing VCFs with `augur index`. This function only reports the strain
names in the given VCF.

We then update the logic of the `filter` module to build a sequence
index when the input is a VCF file and use this to filter the
corresponding metadata. This approach allows us to filter VCF
"sequences" without adding a special VCF-specific filter.