Filter by sequences when they are provided
When sequences are provided, we have filtered by the strains present in
these inputs even if users have not requested sequence outputs. This
simplifies the decision about whether to build a sequence index to the
case when a sequence index isn't provided but sequences are. We still
handle the exception where the `--exclude-all` flag skips the sequence
index building step.