Refine interface for read_metadata
Replace open-ended kwargs with specific keyword arguments for
`read_metadata` and name these arguments (and the corresponding augur
filter CLI arguments) to be less pandas-specific and closer to what we'd
call these variables elsewhere in Augur (e.g., "chunksize" becomes
"chunk_size").

As part of this renaming, this commit also fixes a bug in the dtype
assignment for the pandas CSV reader by only specifying a dtype for the
index column we end up using.

This commit also adds a couple of simple doctests for `read_metadata` to
clarify its expected behavior under common use cases.