Improve helpfulness of the missing id columns error message
Listing both the given id columns and the found metadata columns makes
it easier for someone to debug directly from the error message.

Switches from a KeyError to plain Exception since KeyError values should
be just the key name itself.  Includes the full error message in the
doctest as an example of what to expect (and to make sure that our
interpolations are working as intended).