Skip parsing of missing data in metadata
Fixes an issue with metadata parsing where pandas treats missing
data (e.g., empty strings) as `NaN` values, missing date values (which
appear in some GenBank SARS-CoV-2 records) get interpretted as `float`
values, and downstream functions that expect strings (like
`is_date_ambiguous`) break. This commit resolves the issue by restoring
Augur's original behavior which was to fill all missing values in a data
frame with empty strings [1]. However, we use a slightly different
approach here of asking pandas not to parse missing data at all through
the `na_filter=False` argument. This argument has the same effect as the
previous implementation, but it only needs to be written once to apply
for all calls of `read_metadata`, whereas the `fillna` approach would
need to be applied to data frames or data frame chunks in different
contexts.

To recreate the original issue, this commit updates the functional tests
for augur filter to include a metadata record with a missing date
column and updates the other parts of the functional tests accordingly.

[1] https://github.com/nextstrain/augur/blob/15dfc8b/augur/util_support/metadata_file.py#L97