Test grouping ambiguous dates by month
Adds a functional test for grouping ambiguous dates by year and month
and corresponding metadata, sequence, and sequence index entry for a
strain with only a year in the date field. The test fails because the
ambiguous month gets generated twice for the same strain during the
augur filter command and the group from the first pass through the
metadata doesn't exist in the second pass.

Confirms #760