Track strains skipped during grouping
Updates the interface to the `get_groups_for_subsampling` function such
that it returns a list of strains that have been skipped and the reason
why (following the pattern of filtered and force-included strains lists
from the `apply_filters` function). We count the number of strains
skipped for different reasons, report this number in the final output,
and also optionally log these strains to the output log table. This
change allows users to track why specific strains were excluded from an
analysis after the fact without needing to refer to the logs for stdout
or stderr.