Skip strains with missing months during group by
Instead of randomly generating month values for strains that do not have
any month information, skip these strains and log the reason why they
were skipped. This is a breaking change to the behavior of augur
filter's interface for users who rely on year-only date records to be
grouped anyway. However, this change ensures that grouping results are
more stable and predictable across runs.

Fixes #760