Refactor priorities into functions
This is in preparation for an `augur subsample` command to call these
python functions directly, as opposed to scripts called from the
command line.

The work done by the functions is unchanged, thus they still don't
return any information rather they write information to disk. Future
work will change this so that information can flow back to the calling
program without needing to be written to disk.