Fix issues with reproducible subsampling in filter
Addresses three cases of non-deterministic data processing in the filter
module that caused output for the same command with a random seed to
change between runs. The three main cases are:

1. When adding strains to priority queues, the non-deterministic order
of strains stored in sets causes deterministic "random" priorities for a
given random seed to be assigned to different strains.

2. When creating randomly-sized priority queues by sampling from a
Poisson distribution, that sampling did not use the user-provided random
seed and would produce randomly-sized queues.

3. Groups passed to the `create_queues_by_group` function (used when the
user provides `--subsample-max-sequences`) were randomly ordered (due to
random order of strains stored in sets) causing different groups to get
different sized queues between runs (even when queue size was stable
across runs).

This commit adds unit and functional tests to catch each of these cases
and adds the code changes required for these tests to pass.

Thank you to Philip Shirk for catching this issue and describing it on
the Nextstrain discussion site [1].

[1] https://discussion.nextstrain.org/t/augur-filter-subsample-seed-reproducible-example/723