Fix tests that break with BioPython 1.67
Fixes unit tests that pass a `Path` instance to a BioPython function
that expects a string instead. We now just pass the string.

Also fixes a functional test with the Zika build. Earlier versions of
BioPython (1.67) include trailing punctuation in sequence ids from
GenBank records that later version apparently truncate. For example, our
functional test for Zika includes a reference GenBank record for which
the id ends in a ".". When I run the cram tests for the Zika
build (`zika.t`) with BioPython 1.67, the actual output conflicts with
the expected output (generated from a later BioPython) where there was
no trailing ".", as shown in the diff output below:

```
+  ->KX369547.1 Zika virus strain PF13/251013-18, complete genome
+  +>KX369547.1 Zika virus strain PF13/251013-18, complete genome.
```

This commit specifically ignores the reference line in the diff between
actual and expected alignments, to maintain backward compatibility with
BioPython 1.67.