Create MutableSeq objects with strings
Older versions of BioPython (1.76 and before) do not support passing
`Seq` objects to initialize a `MutableSeq` object, although BioPython
1.79 does support this behavior. In 1.76, this initialization causes a
`TypeError` to be thrown immediately. In 1.67, the initialization
appears to work as expected, but attempts to modify the `MutableSeq`
instance produce a `TypeError` saying that `Seq` objects do not support
item assignment.

This commit resolves these issues for older versions of BioPython that
we still support, by passing the sequence as a string to `MutableSeq`.
This approach works for all support BioPython versions.

Fixes #798