schemas: Add "$id" key so we have a way of referring to these schemas
Useful to unambiguously reference in documentation or even in other
schemas.

While these URLs are not required to be fetchable, I do plan to make
them so by setting up redirects from nextstrain.org to the files on
GitHub.  The reason for this indirection is that nextstrain.org lets us
more easily move the files around (e.g. renaming or to S3) if we ever
need to while preserving old URLs.  This is important as ids should be
~permanent since they're consumed and used externally.  The indirection
will also let us present rendered, human-friendly versions of the
schemas to browsers/humans while still returning the JSON representation
to programmatic clients.