Implicitly group by year and month for month group
Instead of calculating a new (year, month) tuple when users group by
month, add a "year" key to the list of group fields. This fixes a pandas
indexing bug where calling `nlargest` on a SeriesGroupBy object that has
a year and month tuple key for month causes pandas to think the single
month key is a MultiIndex that should be a list. Although this commit is
motivated to fix this pandas issue, this implementation of the
year/month disambiguation is simpler and a more idiomatic pandas
implementation that wouldn't have been possible in the original augur
filter implementation (before we switched to pandas for metadata
parsing).