Separate field titles from groupings
Gives the option to provide display titles for all fields, not just the
groupings. Use separate `groupings` and `filters` arrays to list the
field keys for the group by options and filter options.

Providing `filters` is optional, but it can be used to slim down the
number of options in the filters dropdown if not all fields should be
filters.