Skip codecov and deploy on CI workflows in fork repos (#834)
With #796, our CI workflow can now be run on forks of the repo. I'm currently doing development in a fork victorlin/augur, including for this PR, and CI workflows have been failing due to covtest upload failure albeit passing tests.

This PR adds an extra condition on running the codecov upload and deploy step, only under the context of the main GitHub repository.