Set default filter priority to negative infinity
Given that user-provided priorities to augur filter can be all negative
values of which some rank higher than others (as in the ncov workflow's
proximity-based prioritization), we cannot use a default priority value
of 0 or we will always prioritize strains that do not have a defined
priority. This commit changes the expected default for priorities to
negative infinity (as reflected in the changed unit test) and updates
the defaultdict factory in `read_priority_scores` to return negative
infinity for missing strains.