Allow users to override all tree builder arguments
Changes the meaning of `--tree-builder-args` from an augmentation of the
default arguments to a full replacement of those arguments. Specifies
the default arguments for each tree builder as a global string that a)
the corresponding tree builder uses when the user does not provide any
overriding arguments and b) the help string displays for each tree
builder so users know what they are overriding. As part of this change,
the tree builder arguments default to the more idiomatic `None` instead
of an empty string.

These changes allow users to substantially alter how specific tree
builders run. For example, users can now enable "ultrafast bootstrap"
mode for IQ-TREE with the `-bb` argument. This argument is mutually
exclusive with the default `-fast` options, so users could not get
bootstrap values before now.

Fixes #778
Fixes #780