Check for conflicts with default and user args
Instead of allowing users to provide tree builder arguments that could
conflict with the hardcoded defaults (alignment input, number of
threads, and substitution_model model), check for these conflicts and
report them to the user as an error. This behavior should avoid
surprising results from the different tree builders which appear to be
tolerant of duplicate arguments and yet still produce unexpected errors.