Allow users to optionally override default args
Adds a new flag to augur tree, `--override-default-args`, that allows
users to override default arguments for tree builders with the values
they provide in `--tree-builder-args`. This flag allows us to maintain
the original behavior of `--tree-builder-args`, avoiding breaking
numerous workflows that depend on it. It also allows us to opt in to
behavior required for bootstrap analyses with IQ-TREE, etc.

Adds functional tests to confirm the original and new behavior with and
without the `--override-default-args` flag, respectively.