schemas: version property for dataset v2 is a constant
I think the pattern originally came out of the idea that this schema was
a proposal encompassing not just the next version (v2) but future
versions.  There may also have been some idea that a single schema would
be used for several versions.  In practice, our versions designate major
compatibility levels, so we'll copy + bump for v3.