schemas: Add frequencies output and tip-frequencies sidecar schemas
Useful to have a description of the formats for reference.  These can be
used and extended in the future for generation of documentation pages.

The output of `augur frequencies` can take several forms, one of which
is the tip-frequencies sidecar form (technically a very slight subset of
it).

The tip-frequencies sidecar supported by Auspice takes a specific form
and supports some properties not generated by Augur but in use with
custom generators.  It may also be useful for file format detection in
`nextstrain remote upload`.