[export] order keys in dataset for easier reading
When scanning through dataset JSONs it's much nicer to have certain
keys come first in each dict. For instance, the title of the dataset,
or the name of each tree node. It's especially handy to encode the
data for each node before the children nodes (and this becomes worse
the deeper the nesting!). This has been a personal wish of mine for a
long time.

This commit changes the `write_json` util function to sort keys (the
previous default) unless the data is an OrderedDict (or subclass of).
A search through the augur code indicates that we don't use OrderedDicts
explicitly, however we may be using them implicitly.

All tests pass.