Update link to pandas DataFrame query doc
The section ID on the docs is now `the-query-method`. The ID `indexing-query` is on an invisible `span` element. It was probably an older section ID that's kept around on the span to make sure existing links still work. Updating the link to the new ID selector in case the empty span gets removed in the future.