Add tests for the behavior we expect for fix_dates
Updates functional tests for the parse command to use the `--fix-dates`
argument and updates the corresponding initial sequence data to include
a sequence with an incomplete date value that needs to be fixed. This
functional test fails, recapitulating issue #873.

Also, updates the unit tests for the `fix_dates` function to explicitly
check for warning messages when users provide invalid dates. These tests
now include an example where a user may provide a valid ambiguous date
that currently produces a warning and shouldn't.