Omit duplicate sequences in filter output
Instead of dutifully writing all sequences for strains that pass given
filters, only write out the first sequence for each strain from the
given input sequences and print a warning to stderr when duplicates are
detected. This change allows Augur and the workflows that depend on it
to be more robust to upstream data quality issues.