measurements: use global dict to hold default args
A subset of optional args (currently just `title` and `x-axis-label`)
need a default value for Auspice but we also do not want to overwrite
the config file with the argument default values every time. Use a
global dict `DEFAULT_ARGS` to hold these default values instead of the
argparse `default` argument.