Explicitly close metadata to avoid unclosed file warnings
Splits up the first load of the metadata (where we check for available
id columns) into separate variables for metadata and the first chunk and
then explicitly closes the metadata file after the read. This close call
is required to prevent seemingly random "unclosed file" warnings from
appearing in stderr.