Print helpful error when metadata has invalid id
Adds a functional test for how we'd like augur export v2 to behave when
the user provides metadata without a valid id and then updates the
try/except block in export_v2 to catch and print the error from
`read_metadata` when it occurs.

While we're at it, this commit also prints the error for missing
metadata file to stderr instead of stdout.