Remove option to publish to TestPyPI
Running this workflow against TestPyPI followed by PyPI does not work.

Publishing to TestPyPI was desirable initially when testing this out, but we ran into some unintended consequences when publishing to TestPyPI during an actual release.

On top of publishing as a final step, this workflow also does creates a git commit on master+release and creates a new tag, regardless of which PyPI instance is targeted. So, by running the 15.0.2 release against TestPyPI, the tag is created and a subsequent 15.0.2 release against PyPI would fail.

We have already released to PyPI using this so there is not much benefit keeping TestPyPI around anymore.