Remove usage of release branch
The release process can function fine without a release branch, and he version tags should be sufficient for anyone looking for the latest release in git history.

If anything, we can replace the release branch with a git tag which is updated by the release script, but I don't see any immediate need for one.