measurements export: add `--include-columns` option
Allows users to specify a list of columns from their collection TSV to
include in the output measurements JSON. All columns will be included by
default if option is not provided.

The 'strain' and 'value' columns do not have to be included in the list
since these are required columns. However, other configuration columns
(e.g. groupings) will need to be explicitly included in the list if
users are using the option. This ensures that we do not give unexpected
outputs by auto-including grouping columns.

Includes new functional test for the new option.